+++ cmd/aws/s3delete.go
