+++ pkg/gdalutils.go
